Consolidate ChildListContext into EditorContext.
