+++ client/components/edit/ObjectEditor.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect } from "react";
 import ChildList from "./children/ChildList";
-import { ChildListContextProvider } from "../../context/ChildListContext";
 import Breadcrumbs from "./Breadcrumbs";
 import ObjectSummary from "./ObjectSummary";
 import Box from "@mui/material/Box";
@@ -51,9 +50,7 @@ const ObjectEditor = ({ pid }: ObjectEditorProps): React.ReactElement => {
 
             <DatastreamModal />
             <h2>Contents</h2>
-            <ChildListContextProvider>
-                <ChildList pid={pid} />
-            </ChildListContextProvider>
+            <ChildList pid={pid} />
             <EditorSnackbar />
         </div>
     );
