+++ client/context/EditorContext.test.tsx
@@ -10,7 +10,7 @@ jest.mock("./FetchContext", () => ({
 }));
 
 describe("useEditorContext", () => {
-    let fetchValues;
+    let fetchValues: Record<string, Record<string, () => unknown>>;
     beforeEach(() => {
         fetchValues =  {
             action: {
@@ -215,6 +215,61 @@ describe("useEditorContext", () => {
                 await result.current.action.loadCurrentObjectDetails();
             });
             expect(result.current.action.extractFirstMetadataValue("field", "default")).toEqual("foo");
+        });
+    });
+
+    describe("getChildListStorageKey", () => {
+        it("generates appropriate keys", async () => {
+            const { result } = await renderHook(() => useEditorContext(), { wrapper: EditorContextProvider });
+            expect(result.current.action.getChildListStorageKey("test:123", 1, 10)).toEqual("test:123_1_10");
+        });
+    });
+
+    describe("loadChildDetailsIntoStorage", () => {
+        it("successfully calls fetch", async () => {
+            const { result } = await renderHook(() => useEditorContext(), { wrapper: EditorContextProvider });
+            expect(Object.keys(result.current.state.childDetailsStorage)).toEqual([]);
+            await act(async () => {
+                await result.current.action.loadChildDetailsIntoStorage("test:123");
+            });
+            expect(Object.keys(result.current.state.childDetailsStorage)).toEqual(["test:123"]);
+            expect(fetchValues.action.fetchJSON).toHaveBeenCalledTimes(1);
+            expect(fetchValues.action.fetchJSON).toHaveBeenCalledWith("http://localhost:9000/api/edit/object/test%3A123/details");
+        });
+
+        it("handles exceptions", async () => {
+            const fetchSpy = jest.spyOn(fetchValues.action, "fetchJSON").mockImplementation(() => { throw new Error("kaboom"); });
+            const consoleSpy = jest.spyOn(console, "error").mockImplementation(jest.fn());
+            const { result } = await renderHook(() => useEditorContext(), { wrapper: EditorContextProvider });
+            await act(async () => {
+                await result.current.action.loadChildDetailsIntoStorage("test:123");
+            });
+            expect(fetchSpy).toHaveBeenCalledTimes(1);
+            expect(consoleSpy).toHaveBeenCalledWith("Problem fetching details from http://localhost:9000/api/edit/object/test%3A123/details");
+        });
+    });
+
+    describe("loadChildrenIntoStorage", () => {
+        it("successfully calls fetch", async () => {
+            const { result } = await renderHook(() => useEditorContext(), { wrapper: EditorContextProvider });
+            expect(Object.keys(result.current.state.childListStorage)).toEqual([]);
+            await act(async () => {
+                await result.current.action.loadChildrenIntoStorage("test:123", 1, 10);
+            });
+            expect(Object.keys(result.current.state.childListStorage)).toEqual(["test:123_1_10"]);
+            expect(fetchValues.action.fetchJSON).toHaveBeenCalledTimes(1);
+            expect(fetchValues.action.fetchJSON).toHaveBeenCalledWith("http://localhost:9000/api/edit/object/test%3A123/children?start=0&rows=10");
+        });
+
+        it("handles exceptions", async () => {
+            const fetchSpy = jest.spyOn(fetchValues.action, "fetchJSON").mockImplementation(() => { throw new Error("kaboom"); });
+            const consoleSpy = jest.spyOn(console, "error").mockImplementation(jest.fn());
+            const { result } = await renderHook(() => useEditorContext(), { wrapper: EditorContextProvider });
+            await act(async () => {
+                await result.current.action.loadChildrenIntoStorage("test:123", 1, 10);
+            });
+            expect(fetchSpy).toHaveBeenCalledTimes(1);
+            expect(consoleSpy).toHaveBeenCalledWith("Problem fetching tree data from http://localhost:9000/api/edit/object/test%3A123/children?start=0&rows=10");
         });
     });
 });
